+++ controllers/vehicles.js
@@ -1,28 +1,29 @@
 const vehiclesRouter = require('express').Router()
 const Vehicle = require('../models/vehicle')
 const Driver = require('../models/driver')
+const jwt = require('jsonwebtoken')
 
-vehiclesRouter.get('/', async (request, response) => {
+vehiclesRouter.get('/', async (request, response, next) => {
     try {
-        let vehicles = await Vehicle.find({}).populate('driver', {name:1})
-       if(request.query.driver){
-           vehicles = vehicles.filter(v => v.driver.name.toLowerCase() === request.query.driver.toLowerCase() )
-       }
-       if(request.query.class){
-           vehicles = vehicles.filter(v => v.class.toLowerCase() === request.query.class.toLowerCase())
-       }
-       if(request.query.name){
-           vehicles = vehicles.filter(v => v.name.toLowerCase() === request.query.name.toLowerCase())
-       }
+        let vehicles = await Vehicle.find({}).populate('driver', { name: 1 })
+        if (request.query.driver) {
+            vehicles = vehicles.filter(v => v.driver.name.toLowerCase() === request.query.driver.toLowerCase())
+        }
+        if (request.query.class) {
+            vehicles = vehicles.filter(v => v.class.toLowerCase() === request.query.class.toLowerCase())
+        }
+        if (request.query.name) {
+            vehicles = vehicles.filter(v => v.name.toLowerCase() === request.query.name.toLowerCase())
+        }
         response.json(vehicles.map(v => v.toJSON()))
     } catch (exception) {
-        console.log(exception)
+        next(exception)
     }
 })
 
-vehiclesRouter.post('/', async (request, response) => {
+vehiclesRouter.post('/', async (request, response, next) => {
     const body = request.body
-    if(body.driver === undefined || body.name === undefined){
+    if (body.driver === undefined || body.name === undefined) {
         return response.status(400).json({ error: 'some properties are missing' })
     }
     try {
@@ -39,20 +40,39 @@ vehiclesRouter.post('/', async (request, response) => {
         await driver.save()
         response.json(savedVehicle.toJSON())
     } catch (exception) {
-        console.log(expection)
+        next(exception)
     }
 })
 
-vehiclesRouter.delete('/:id', async (request, response) => {
-    try{
+vehiclesRouter.delete('/:id', async (request, response, next) => {
+    const token = request.token
+    if (!token) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const decodedToken = jwt.verify(token, "ASD")
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    try {
         await Vehicle.findByIdAndRemove(request.params.id)
-    }catch(exception){
-        console.log(expection)
+        response.status(200).end()
+    } catch (exception) {
+        next(exception)
     }
 })
 
-vehiclesRouter.put('/:id', async (request, response) => {
+vehiclesRouter.put('/:id', async (request, response, next) => {
     const body = request.body
+    const token = request.token
+    if (!token) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const decodedToken = jwt.verify(token, "ASD")
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
     const vehicleToUpdate = await Vehicle.findById(request.params.id)
 
     const vehicle = {
@@ -61,12 +81,12 @@ vehiclesRouter.put('/:id', async (request, response) => {
         name: body.name === undefined ? vehicleToUpdate.name : body.name,
         description: body.description === undefined ? vehicleToUpdate.description : body.description
     }
-    
-    try{
-        const updatedVehicle = await Vehicle.findByIdAndUpdate(request.params.id,vehicle, {new: true})
+
+    try {
+        const updatedVehicle = await Vehicle.findByIdAndUpdate(request.params.id, vehicle, { new: true })
         response.json(updatedVehicle.toJSON())
-    }catch(exception){
-        console.log(expection)
+    } catch (exception) {
+        next(exception)
     }
 })
